feat(widget): add split chunks to build
