+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@pontem/liquidswap-widget",
-  "version": "0.1.5",
+  "version": "0.1.6",
   "homepage": "https://github.com/pontem-network/liquidswap-widget#readme",
   "files": [
     "src",
@@ -36,23 +36,21 @@
     "deploy": "yarn build:gh-pages; gh-pages -d docs"
   },
   "devDependencies": {
-    "@esbuild-plugins/node-globals-polyfill": "^0.1.1",
-    "@esbuild-plugins/node-modules-polyfill": "^0.1.4",
     "@originjs/vite-plugin-commonjs": "^1.0.3",
     "@types/webpack-env": "^1.18.0",
     "@vitejs/plugin-vue": "^4.0.0",
     "gh-pages": "^4.0.0",
-    "rollup-plugin-node-polyfills": "^0.2.1",
+    "vite-plugin-chunk-split": "^0.4.7",
     "vite-plugin-dts": "^1.7.1",
-    "vue-tsc": "^1.0.11",
-    "vite-svg-loader": "4.0.0"
+    "vite-svg-loader": "4.0.0",
+    "vue-tsc": "^1.0.11"
   },
   "dependencies": {
     "@metamask/jazzicon": "2.0.0",
-    "@tsconfig/recommended": "^1.0.1",
     "@pontem/aptos-wallet-adapter": "^0.6.8",
     "@pontem/coins-registry": "^1.0.4",
     "@pontem/liquidswap-sdk": "^0.5.4",
+    "@tsconfig/recommended": "^1.0.1",
     "@types/lodash": "4.14.182",
     "@vueuse/core": "^8.9.4",
     "@vueuse/shared": "^9.9.0",
